Move Asset class from View to Routing.
